requirements.txt
torch==2.2.2
torchaudio
transformers
datasets<=4.5.0
jiwer
numpy<2
openai
python-dotenv
